[refactor #29] MemberException 정리
