+++ src/main/java/com/project/dogwalker/exception/GlobalExceptionHandler.java
@@ -8,11 +8,8 @@ import com.project.dogwalker.exception.feign.FeignCommonException;
 import com.project.dogwalker.exception.feign.FeignErrorParseException;
 import com.project.dogwalker.exception.feign.FeignNotFoundException;
 import com.project.dogwalker.exception.feign.FeignServerException;
-import com.project.dogwalker.exception.member.ImgUploadFailException;
-import com.project.dogwalker.exception.member.LoginMemberNotFoundException;
-import com.project.dogwalker.exception.member.MemberNotFoundException;
-import com.project.dogwalker.exception.member.NotWalkerException;
-import com.project.dogwalker.exception.member.WalkerNotWritePriceException;
+import com.project.dogwalker.exception.member.AuthMemberException;
+import com.project.dogwalker.exception.member.MemberException;
 import com.project.dogwalker.exception.notice.NoticeNotFoundException;
 import com.project.dogwalker.exception.notice.SseException;
 import com.project.dogwalker.exception.reserve.AlreadyUnLockException;
@@ -74,30 +71,17 @@ public class GlobalExceptionHandler {
     return ResponseEntity.internalServerError().body(ExceptionResponse.from(e));
   }
 
-  @ExceptionHandler(WalkerNotWritePriceException.class)
-  public ResponseEntity<ExceptionResponse> handleWalkerNotWritePRice(final WalkerNotWritePriceException e){
-    log.info(LOG_ERROR_MESSAGE,e.getClass(),e.getErrorCode(),e.getMessage());
-    return ResponseEntity.badRequest().body(ExceptionResponse.from(e));
-  }
-
-  @ExceptionHandler(LoginMemberNotFoundException.class)
-  public ResponseEntity<TokenResponse> hanldeNotFoundMember(final LoginMemberNotFoundException e){
+  @ExceptionHandler(AuthMemberException.class)
+  public ResponseEntity<TokenResponse> hanldeAuthException(final AuthMemberException e){
     log.info(LOG_ERROR_MESSAGE,e.getClass(),e.getErrorCode(),e.getMessage());
     return ResponseEntity.status(HttpStatus.NOT_FOUND).body(TokenResponse.from(e,e.getToken()));
   }
 
-  @ExceptionHandler(MemberNotFoundException.class)
-  public ResponseEntity<ExceptionResponse> handleMemberNotFound(final MemberNotFoundException e){
+  @ExceptionHandler(MemberException.class)
+  public ResponseEntity<ExceptionResponse> handleMemberException(final MemberException e){
     log.info(LOG_ERROR_MESSAGE,e.getClass(),e.getErrorCode(),e.getMessage());
     return ResponseEntity.badRequest().body(ExceptionResponse.from(e));
   }
-
-  @ExceptionHandler(NotWalkerException.class)
-  public ResponseEntity<ExceptionResponse> handleNotWalker(final NotWalkerException e){
-    log.info(LOG_ERROR_MESSAGE,e.getClass(),e.getErrorCode(),e.getMessage());
-    return ResponseEntity.badRequest().body(ExceptionResponse.from(e));
-  }
-
 
   @ExceptionHandler(AlreadyUnLockException.class)
   public ResponseEntity<ExceptionResponse> handleAlreadyUnLock(final AlreadyUnLockException e){
