+++ src/main/java/com/project/dogwalker/exception/member/AuthMemberException.java
@@ -5,9 +5,10 @@ import com.project.dogwalker.exception.ErrorCode;
 import lombok.Getter;
 
 @Getter
-public class NotWalkerException extends CustomException {
-
-  public NotWalkerException(final ErrorCode errorCode ) {
+public class AuthMemberException extends CustomException {
+  private String token;
+  public AuthMemberException(final ErrorCode errorCode ,final String token) {
     super(errorCode);
+    this.token = token;
   }
 }
