+++ src/main/java/com/project/dogwalker/walkersearch/service/WalkerInfoServiceImpl.java
@@ -7,7 +7,7 @@ import com.project.dogwalker.domain.user.User;
 import com.project.dogwalker.domain.user.UserRepository;
 import com.project.dogwalker.domain.user.walker.elastic.WalkerDocument;
 import com.project.dogwalker.domain.user.walker.elastic.WalkerSearchRepository;
-import com.project.dogwalker.exception.member.MemberNotFoundException;
+import com.project.dogwalker.exception.member.MemberException;
 import com.project.dogwalker.member.dto.MemberInfo;
 import com.project.dogwalker.walkersearch.dto.WalkerInfo;
 import com.project.dogwalker.walkersearch.dto.WalkerInfoSearchCond;
@@ -39,7 +39,7 @@ public class WalkerInfoServiceImpl implements WalkerInfoService {
   @Transactional(readOnly = true)
   public List <WalkerInfo> getWalkerInfoList(final MemberInfo info ,final WalkerInfoSearchCond searchCond) {
     final User user = userRepository.findByUserEmailAndUserRole(info.getEmail() , info.getRole())
-        .orElseThrow(() -> new MemberNotFoundException(NOT_EXIST_MEMBER));
+        .orElseThrow(() -> new MemberException(NOT_EXIST_MEMBER));
     if(searchCond.getLat()==null||searchCond.getLnt()==null){
       searchCond.setLat(user.getUserLat());
       searchCond.setLnt(user.getUserLnt());
@@ -58,7 +58,7 @@ public class WalkerInfoServiceImpl implements WalkerInfoService {
   @Transactional(readOnly = true)
   public WalkerUnAvailDetail getWalkerUnAvailService(Long walkerId) {
     final User walker = userRepository.findByUserIdAndUserRole(walkerId , WALKER)
-        .orElseThrow(() -> new MemberNotFoundException(NOT_EXIST_MEMBER));
+        .orElseThrow(() -> new MemberException(NOT_EXIST_MEMBER));
 
     final List <WalkerPermUnAvailDate> walkerPermUnAvailDates = userRepository.walkerPermUnVailScheduleFindByWalkerId(
         walker.getUserId());
